+++ src/components/Capture/capture.js
@@ -10,19 +10,17 @@ import { canvasToBase64Images } from '../utils/canvas.js'
 import { base64toBlob, fileToBase64, isOfFileType, fileToLossyBase64Image } from '../utils/file.js'
 import { postToBackend } from '../utils/sdkBackend'
 
-let hasWebcamStartupValue = true;//asume there is a webcam first,
-//assuming it's better to get flicker from webcam to file upload
-//than the other way around
-
-checkIfHasWebcam( hasWebcam => hasWebcamStartupValue = hasWebcam )
-
 class Capture extends Component {
+  static defaultProps = {
+    onWebcamSupportChange: () => {},
+  }
+
   constructor (props) {
     super(props)
     this.state = {
       uploadFallback: false,
       error: null,
-      hasWebcam: hasWebcamStartupValue
+      hasWebcam: undefined,
     }
   }
 
@@ -31,11 +29,15 @@ class Capture extends Component {
     this.checkWebcamSupport();
   }
 
+  componentWillUpdate(nextProps, nextState) {
+    if (nextState.hasWebcam !== this.state.hasWebcam) {
+      this.props.onWebcamSupportChange(nextState.hasWebcam)
+    }
+  }
+
   componentWillUnmount () {
     this.setState({uploadFallback: false})
-    clearInterval(this.webcamChecker);
-    // TODO add this callback when implementing live capture
-    // useFullScreen(false)
+    clearInterval(this.webcamChecker)
   }
 
   componentWillReceiveProps(nextProps) {
@@ -59,6 +61,12 @@ class Capture extends Component {
     const { actions, method, side } = this.props
     const capture = {...payload, side}
     actions.createCapture({method, capture, maxCaptures: this.maxAutomaticCaptures})
+  }
+
+  createLivenessVideo(isLiveness, url) {
+    const payload = {isLiveness, url}
+    this.createCapture(payload)
+    this.validateAndProceed(payload)
   }
 
   validateAndProceed(payload) {
@@ -85,6 +93,11 @@ class Capture extends Component {
       document: () => this.handleDocument(payload),
       face: () => this.handleFace(payload)
     })
+  }
+
+  onVideoRecorded = (blob) => {
+    const url = window.URL.createObjectURL(blob);
+    this.createLivenessVideo(this.props.liveness, url)
   }
 
   initialiseCapturePayload = (blob, base64) => ({id: randomId(), blob, base64})
@@ -136,7 +149,6 @@ class Capture extends Component {
   }
 
   onWebcamError = () => {
-    this.setState({uploadFallback: true})
     this.deleteCaptures()
   }
 
@@ -202,25 +214,27 @@ class Capture extends Component {
     this.setState({error: null})
   }
 
-  render ({useWebcam, back, i18n, termsAccepted, ...other}) {
-    const useCapture = (!this.state.uploadFallback && useWebcam && isDesktop && this.state.hasWebcam)
-    // TODO add this callback when implementing live capture
-    // useFullScreen(useCapture)
+  render ({useWebcam, back, i18n, termsAccepted, liveness, ...other}) {
+    const canUseWebcam = this.state.hasWebcam && !this.state.uploadFallback
+    const shouldUseWebcam = useWebcam && (this.props.method === 'face' || isDesktop)
+    const useCapture = canUseWebcam && shouldUseWebcam
+
     return (
       process.env.PRIVACY_FEATURE_ENABLED && !termsAccepted ?
         <PrivacyStatement {...{i18n, back, acceptTerms: this.acceptTerms, ...other}}/> :
-        <CaptureMode {...{useCapture, i18n,
+        <CaptureMode {...{useCapture, liveness, i18n,
           onScreenshot: this.onScreenshot,
+          onVideoRecorded: this.onVideoRecorded,
           onUploadFallback: this.onUploadFallback,
           onImageSelected: this.onImageFileSelected,
-          onWebcamError: this.onWebcamError,
+          onFailure: this.onWebcamError,
           error: this.state.error,
           ...other}}/>
     )
   }
 }
 
-const CaptureMode = ({method, documentType, side, useCapture, i18n, ...other}) => {
+const CaptureMode = ({method, documentType, side, useCapture, i18n, liveness, ...other}) => {
   const copyNamespace = method === 'face' ? 'capture.face' : `capture.${documentType}.${side}`
   const title = !useCapture && i18n.t(`${copyNamespace}.upload_title`) ? i18n.t(`${copyNamespace}.upload_title`)  : i18n.t(`${copyNamespace}.title`)
   const subTitle = useCapture && isDesktop ? i18n.t(`${copyNamespace}.webcam`) : null
@@ -228,9 +242,9 @@ const CaptureMode = ({method, documentType, side, useCapture, i18n, ...other}) =
   const parentheses = i18n.t('capture_parentheses')
   return (
     useCapture ?
-      <Camera {...{i18n, method, title, subTitle, ...other}}/> :
+      <Camera {...{i18n, method, title, subTitle, liveness, ...other}}/> :
       <Uploader {...{i18n, instructions, parentheses, title, subTitle, ...other}}/>
-  )
+    )
 }
 
 const mapStateToProps = (state, props) => {