+++ src/components/Countdown/index.js
@@ -1,6 +1,0 @@
-import { h } from 'preact'
-import style from './style.css'
-
-const Countdown = () => (<span className={style.countdown}></span>)
-
-export default Countdown