+++ test/features/helpers/i18n_helper.rb
@@ -1,12 +1,10 @@
 class I18nHelper
   def load_locale(locale)
     tag = locale ? locale : 'en'
-    lines = File.readlines(File.expand_path("../src/locales/#{tag}.js"))
-    lines[0] = "{\n"
-    @translations = lines.join.to_json
+    @translations = JSON.parse(File.read(File.expand_path("../src/locales/#{tag}.json")))
   end
 
   def translate(key)
-    @translations[key]
+    @translations.dig(*key.split(".")) or key
   end
 end