+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This change log file is based on best practices from [Keep a Changelog](http://k
 This project adheres to [Semantic Versioning](http://semver.org/). Breaking changes result in a different MAJOR version. UI changes that might break customizations on top of the SDK will be treated as breaking changes too.
 This project adheres to the Node [default version scheme](https://docs.npmjs.com/misc/semver).
 
+## [2.5.0] - 2018-07-27
+
+### Added
+- UI: Added a permission priming screen to inform the user that camera permissions must be enabled.
+- UI: Added a permission recovering screen in case user media permissions are denied.
+- UI: Added intro screen when entering cross device flow.
+
+### Changed
+- UI: Changed UI for face capture step. On small screens it will display a full screen camera component.
+- UI: Desktop - If webcam is not available, a cross device intro screen will be shown to allow the user to take a live photo on their mobile phones.
+
 ## [2.4.1] - 2018-05-18
 
 ### Fixed
@@ -267,7 +278,8 @@ Install with `npm install onfido-sdk-ui@0.12.0-rc.1`
 - NPM (commonjs2) style of importing the library now works
 
 
-[next-version]: https://github.com/onfido/onfido-sdk-ui/compare/2.4.1...development
+[next-version]: https://github.com/onfido/onfido-sdk-ui/compare/2.5.0...development
+[2.5.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.4.1...2.5.0
 [2.4.1]: https://github.com/onfido/onfido-sdk-ui/compare/2.4.0...2.4.1
 [2.4.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.2.0...2.3.0